add scrub routines
