+++ lib/vger_vendor_loads.rb
@@ -1,1 +1,2 @@
 require_relative './vger_vendor_loads/loader'
+require_relative './vger_vendor_loads/fixes'
